Change code structure
While there was nothing wrong with the code at all, I just felt a bit of
cleaning would ease readability. This is really subjective changes and
should only be taken as my opinion of how to structure ruby code.

I do not want to step on any toes, so please let me know if this is
overstepping.